# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.ExportBank:
    properties:
      bank_type:
        example: theory
        type: string
      difficulty:
        example: medium
        type: string
      language:
        example: go
        type: string
      questions:
        items:
          $ref: '#/definitions/api.ExportQuestion'
        type: array
      subject:
        example: Go concurrency patterns
        type: string
    type: object
  api.ExportCategory:
    properties:
      banks:
        items:
          $ref: '#/definitions/api.ExportBank'
        type: array
      name:
        example: Golang
        type: string
    type: object
  api.ExportData:
    properties:
      categories:
        description: Categories without a folder
        items:
          $ref: '#/definitions/api.ExportCategory'
        type: array
      exported_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      folders:
        items:
          $ref: '#/definitions/api.ExportFolder'
        type: array
      version:
        example: "1.1"
        type: string
    type: object
  api.ExportFolder:
    properties:
      categories:
        items:
          $ref: '#/definitions/api.ExportCategory'
        type: array
      name:
        example: Programming
        type: string
    type: object
  api.ExportQuestion:
    properties:
      expected_answer:
        example: A goroutine is a lightweight thread managed by the Go runtime.
        type: string
      grading_prompt:
        type: string
      hint:
        type: string
      subject:
        example: What is a goroutine?
        type: string
    type: object
...
  api.ImportResult:
    properties:
      banks_created:
        example: 5
        type: integer
      categories_created:
        example: 2
        type: integer
      errors:
        description: |-
          Errors lists everything that was skipped. Omitted entirely when the import
          was clean, so its presence alone means "not everything came through".
          The import is not transactional — see importAll — so a partial result is a
          real outcome the user has to be told about rather than a failure to hide.
        items:
          type: string
        type: array
      folders_created:
        example: 3
        type: integer
      questions_created:
        example: 42
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal learning companion — create question banks, practice, and
    let AI grade your answers.
  title: Remaimber-it API
  version: "1.0"
paths:
...
  /export:
    get:
      description: Export all folders, categories, banks, and questions as a downloadable
        JSON file. The system "Deleted" folder and its contents are excluded.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ExportData'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export all data
      tags:
      - Import/Export
...
  /import:
    post:
      consumes:
      - application/json
      description: Import folders, categories, banks, and questions from a JSON export.
        New IDs are generated for all entities.
      parameters:
      - description: Export data to import
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ExportData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.ImportResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Import data
      tags:
      - Import/Export
...
swagger: "2.0"